+++ src/snw/engine/component/SetSwitch.java
@@ -102,7 +102,7 @@ public class SetSwitch extends FrameComponent {
     }
 
     @Override
-    public void mouseClicked(int mouseX, int mouseY) {
+    public void mouseClicked(double mouseX, double mouseY) {
         if (mouseX < getWidth() / 2) {
             switchLeft();
         } else {
